backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file backend/requirements.txt
argcomplete==3.6.2
    # via commitizen
asgiref==3.8.1
    # via django
black==25.1.0
    # via swiss (pyproject.toml)
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via commitizen
click==8.2.1
    # via black
colorama==0.4.6
    # via commitizen
commitizen==4.8.3
    # via swiss (pyproject.toml)
decli==0.6.3
    # via commitizen
distlib==0.3.9
    # via virtualenv
django==5.2.3
    # via
    #   swiss (pyproject.toml)
    #   django-filter
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==25.1
    # via swiss (pyproject.toml)
django-phonenumber-field==8.1.0
    # via swiss (pyproject.toml)
djangorestframework==3.16.0
    # via
    #   swiss (pyproject.toml)
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via swiss (pyproject.toml)
filelock==3.18.0
    # via virtualenv
flake8==7.2.0
    # via pytest-flake8
identify==2.6.12
    # via pre-commit
importlib-metadata==8.7.0
    # via commitizen
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via swiss (pyproject.toml)
jinja2==3.1.6
    # via commitizen
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   commitizen
    #   pytest
pathspec==0.12.1
    # via black
phonenumbers==9.0.7
    # via django-phonenumber-field
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via swiss (pyproject.toml)
prompt-toolkit==3.0.51
    # via questionary
psycopg==3.2.9
    # via swiss (pyproject.toml)
pycodestyle==2.13.0
    # via flake8
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via pytest
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytest==8.4.0
    # via
    #   pytest-django
    #   pytest-flake8
pytest-django==4.7.0
    # via swiss (pyproject.toml)
pytest-flake8==1.3.0
    # via swiss (pyproject.toml)
python-dotenv==1.1.0
    # via swiss (pyproject.toml)
pytz==2025.2
    # via swiss (pyproject.toml)
pyyaml==6.0.2
    # via
    #   commitizen
    #   pre-commit
questionary==2.1.0
    # via commitizen
sqlparse==0.5.3
    # via django
termcolor==3.1.0
    # via commitizen
tomlkit==0.13.3
    # via commitizen
virtualenv==20.31.2
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
zipp==3.23.0
    # via importlib-metadata
